+++ crates/holochain/src/core/workflow/app_validation_workflow/run_validation_callback_tests.rs
@@ -30,7 +30,7 @@ use holochain_types::{
 use holochain_wasm_test_utils::TestWasm;
 use holochain_wasmer_host::module::ModuleCache;
 use holochain_zome_types::{
-    chain::{ChainFilter, LimitConditions, MustGetAgentActivityInput},
+    chain::{ChainFilter, MustGetAgentActivityInput},
     dependencies::holochain_integrity_types::{UnresolvedDependencies, ValidateCallbackResult},
     entry::MustGetActionInput,
     fixt::{CreateFixturator, DeleteFixturator, SignatureFixturator},
@@ -42,7 +42,7 @@ use holochain_zome_types::{
 };
 use matches::assert_matches;
 use parking_lot::RwLock;
-use std::{collections::HashSet, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 
 // test app validation with a must get action where the original action of
 // a delete is not in the cache db and then added to it
@@ -231,13 +231,10 @@ async fn validation_callback_awaiting_deps_agent_activity() {
         move |api, op: Op| {
             if let Op::RegisterDelete(RegisterDelete { delete }) = op {
                 // chain filter with delete as chain top and create as chain bottom
-                let mut filter_hashes = HashSet::new();
-                filter_hashes.insert(delete.hashed.deletes_address.clone().clone());
-                let chain_filter = ChainFilter {
-                    chain_top: delete.as_hash().clone(),
-                    limit_conditions: LimitConditions::UntilHash(filter_hashes),
-                    include_cached_entries: false,
-                };
+                let chain_filter = ChainFilter::until_hash(
+                    delete.as_hash().clone(),
+                    delete.hashed.deletes_address.clone(),
+                );
                 let result = api.must_get_agent_activity(MustGetAgentActivityInput {
                     author: delete.hashed.author.clone(),
                     chain_filter: chain_filter.clone(),
@@ -298,11 +295,13 @@ async fn validation_callback_awaiting_deps_agent_activity() {
     // return single action as requested chain
     network.expect_must_get_agent_activity().returning({
         let expected_chain_top = expected_chain_top.clone();
+        let expected_until_hash = delete.deletes_address.clone();
         let create_action_signed_hashed = create_action_signed_hashed.clone();
         let delete_action_signed_hashed = delete_action_signed_hashed.clone();
         move |author, filter, _, _| {
             assert_eq!(author, alice);
             assert_eq!(&filter.chain_top, expected_chain_top.as_hash());
+            assert_eq!(filter.get_until_hash(), Some(&expected_until_hash));
 
             Ok(vec![MustGetAgentActivityResponse::activity(vec![
                 RegisterAgentActivity {