+++ crates/holochain_cascade/src/error.rs
@@ -27,6 +27,9 @@ pub enum CascadeError {
     #[error("Got an invalid response from an authority for the request hash: {0:?}")]
     InvalidResponse(AnyDhtHash),
 
+    #[error("Input parameters are invalid: {0}")]
+    InvalidInput(String),
+
     #[error(transparent)]
     JoinError(#[from] JoinError),
 
@@ -53,6 +56,9 @@ pub enum CascadeError {
 
     #[error(transparent)]
     SyncScratchError(#[from] holochain_state::scratch::SyncScratchError),
+
+    #[error("Network not initialized")]
+    NetworkNotInitialized,
 }
 
 pub type CascadeResult<T> = Result<T, CascadeError>;